Restructurization and model parser.
