+++ greenfield-validation-model/pom.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <artifactId>veraPDF-greenfield-valdation</artifactId>
+        <groupId>org.verapdf</groupId>
+        <version>1.0-SNAPSHOT</version>
+    </parent>
+
+    <artifactId>greenfield-validation-model</artifactId>
+
+    <dependencies>
+
+        <dependency>
+            <groupId>org.verapdf</groupId>
+            <artifactId>pdflib</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.verapdf</groupId>
+            <artifactId>pdf-model</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.verapdf</groupId>
+            <artifactId>core</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>log4j</groupId>
+            <artifactId>log4j</artifactId>
+        </dependency>
+
+    </dependencies>
+
+</project>
