+++ greenfield-validation-model/src/main/java/org/verapdf/model/impl/cos/GFCosString.java
