+++ pom.xml
@@ -10,9 +10,13 @@
         <version>1.0.13</version>
     </parent>
 
-    <groupId>org.verapdf</groupId>
     <artifactId>veraPDF-greenfield-valdation</artifactId>
     <version>1.0-SNAPSHOT</version>
+    <packaging>pom</packaging>
+
+    <modules>
+        <module>greenfield-validation-model</module>
+    </modules>
 
     <repositories>
         <repository>
@@ -36,29 +40,19 @@
     </pluginRepositories>
 
     <properties>
+        <verapdf.pdflib.version>0.0.1-SNAPSHOT</verapdf.pdflib.version>
     </properties>
 
-    <dependencies>
+    <dependencyManagement>
+        <dependencies>
+            <!-- veraPDF Project dependencies -->
+            <dependency>
+                <groupId>org.verapdf</groupId>
+                <artifactId>pdflib</artifactId>
+                <version>${verapdf.pdflib.version}</version>
+            </dependency>
 
-        <dependency>
-            <groupId>org.verapdf</groupId>
-            <artifactId>pdflib</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.verapdf</groupId>
-            <artifactId>pdf-model</artifactId>
-            <version>[0.13.0,0.14.0)</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.verapdf</groupId>
-            <artifactId>core</artifactId>
-            <version>[0.13.0,0.14.0)</version>
-        </dependency>
-
-    </dependencies>
-
+        </dependencies>
+    </dependencyManagement>
 
 </project>
