--- src/main/java/org/verapdf/model/impl/cos/GFCosNumber.java
@@ -1,44 +1,0 @@
-package org.verapdf.model.impl.cos;
-
-import org.verapdf.cos.COSNumber;
-import org.verapdf.model.coslayer.CosNumber;
-
-/**
- * Created by Timur on 4/8/2016.
- */
-public class GFCosNumber extends GFCosObject implements CosNumber {
-
-    private final long longVal;
-    private final double doubleVal;
-
-    protected GFCosNumber(COSNumber number, final String type) {
-        super(number, type);
-        this.longVal = number.getInteger();
-        this.doubleVal = number.getReal();
-    }
-
-    /**
-     * Get the string representing this object
-     */
-    @Override
-    public String getstringValue() {
-        return String.valueOf(this.doubleVal);
-    }
-
-    /**
-     * Get truncated integer value
-     */
-    @Override
-    public Long getintValue() {
-        return Long.valueOf(this.longVal);
-    }
-
-    /**
-     * Get original decimal value
-     */
-    @Override
-    public Double getrealValue() {
-        return Double.valueOf(this.doubleVal);
-    }
-
-}
